Restart db after checkpoint corrupt. & ensure normal operation.
